Revert "Merge branch 'develop' into sarge-build-notif-location"
This reverts commit a594bc3c9f343ceea4fac0e4b2a8d57204a7da98, reversing
changes made to ae1e9add3db4851dce07084788e9a81f78d9b0f4.

+++ soh/soh/Network/Network.cpp
@@ -1,5 +1,6 @@
 #include "Network.h"
 #include <spdlog/spdlog.h>
+#include <libultraship/libultraship.h>
 
 // MARK: - Public
 
@@ -48,7 +49,7 @@ void Network::ProcessOutgoingPackets() {
 
 void Network::SendDataToRemote(const char* payload) {
     SPDLOG_DEBUG("[Network] Sending data: {}", payload);
-    SDLNet_TCP_Send(networkSocket, payload, static_cast<int>(strlen(payload) + 1));
+    SDLNet_TCP_Send(networkSocket, payload, strlen(payload) + 1);
 }
 
 void Network::SendJsonToRemote(nlohmann::json payload) {
